commit 8-12-23
- Gabung LoginActivity dan RegisterActivity
