+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,13 @@
         android:theme="@style/Theme.CapstoneMD"
         tools:targetApi="31">
         <activity
-            android:name=".detail.DetailActivity"
+            android:name=".ui.login.LoginActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ui.register.RegisterActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ui.detail.DetailActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
