+++ app/src/main/java/com/dicoding/capstonemd/ui/chatbot/ChatbotFragment.kt
@@ -1,21 +1,22 @@
-package com.dicoding.capstonemd.profile
+package com.dicoding.capstonemd.ui.chatbot
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.dicoding.capstonemd.databinding.FragmentProfileBinding
+import com.dicoding.capstonemd.R
+import com.dicoding.capstonemd.databinding.FragmentChatbotBinding
 
-class ProfileFragment : Fragment() {
+class ChatbotFragment : Fragment() {
 
-    private lateinit var binding: FragmentProfileBinding
+    private lateinit var binding: FragmentChatbotBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentProfileBinding.inflate(inflater, container, false)
+        binding = FragmentChatbotBinding.inflate(inflater, container, false)
         return binding.root
     }
