+++ app/src/main/java/com/dicoding/capstonemd/ui/detail/DetailActivity.kt
@@ -1,4 +1,4 @@
-package com.dicoding.capstonemd.detail
+package com.dicoding.capstonemd.ui.detail
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
