+++ app/src/main/java/com/dicoding/capstonemd/ui/detail/TabNutritionFragment.kt
@@ -1,4 +1,4 @@
-package com.dicoding.capstonemd.detail
+package com.dicoding.capstonemd.ui.detail
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,12 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.dicoding.capstonemd.R
 
-class TabMapsFragment : Fragment() {
+class TabNutritionFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_tab_maps, container, false)
+
+        return inflater.inflate(R.layout.fragment_tab_nutrition, container, false)
     }
 }
