+++ app/src/main/java/com/dicoding/capstonemd/ui/profile/ProfileFragment.kt
@@ -1,22 +1,21 @@
-package com.dicoding.capstonemd.chatbot
+package com.dicoding.capstonemd.ui.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.dicoding.capstonemd.R
-import com.dicoding.capstonemd.databinding.FragmentChatbotBinding
+import com.dicoding.capstonemd.databinding.FragmentProfileBinding
 
-class ChatbotFragment : Fragment() {
+class ProfileFragment : Fragment() {
 
-    private lateinit var binding: FragmentChatbotBinding
+    private lateinit var binding: FragmentProfileBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentChatbotBinding.inflate(inflater, container, false)
+        binding = FragmentProfileBinding.inflate(inflater, container, false)
         return binding.root
     }
