+++ app/src/main/res/layout/fragment_tab_maps.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".detail.TabMapsFragment">
+    tools:context=".ui.detail.TabMapsFragment">
 
     <TextView
         android:id="@+id/section_label"
