+++ app/src/main/res/layout/fragment_tab_nutrition.xml
@@ -5,7 +5,7 @@
     android:backgroundTint="#FFB100"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".detail.TabNutritionFragment">
+    tools:context=".ui.detail.TabNutritionFragment">
 
     <TextView
         android:id="@+id/detail_carb_text"
