+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -15,9 +15,4 @@
         android:id="@+id/nav_history"
         android:icon="@drawable/baseline_feed_24"
         android:title="@string/title_history" />
-
-    <item
-        android:id="@+id/nav_profile"
-        android:icon="@drawable/baseline_person_24"
-        android:title="@string/title_profile" />
 </menu>
