+++ app/src/main/res/values/strings.xml
@@ -21,4 +21,28 @@
     <string name="homepage_title_text">Welcome, John Doe!</string>
     <string name="homepage_subtitle_text">Mau makan apa hari ini?</string>
     <string name="see_detail">Lihat Detail</string>
+    <string name="title_login">Login</string>
+    <string name="title_logout">Logout</string>
+    <string name="title_activity_main2">MainActivity2</string>
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="nav_header_title">Android Studio</string>
+    <string name="nav_header_subtitle">android.studio@android.com</string>
+    <string name="nav_header_desc">Navigation header</string>
+    <string name="action_settings">Settings</string>
+
+    <string name="menu_home">Home</string>
+    <string name="menu_gallery">Gallery</string>
+    <string name="menu_slideshow">Slideshow</string>
+    <string name="register_intro_text">Lengkapi data dirimu dibawah ini, ya!</string>
+    <string name="register_name_text">Nama</string>
+    <string name="register_email_text">Email</string>
+    <string name="register_password_text">Password</string>
+    <string name="register_sign_up_text">Sign Up</string>
+    <string name="register_account_text">Already have an account?</string>
+    <string name="register_account_login_text">Log in here.</string>
+    <string name="login_intro_text">Sudah siap untuk makan makanan yang sehat?</string>
+    <string name="login_message_text">Masukkan data dirimu dulu, ya!</string>
+    <string name="login_account_text">Don\'t have an account?</string>
+    <string name="login_account_login_text">Register here.</string>
 </resources>
