+++ app/src/main/res/values/themes.xml
@@ -9,4 +9,13 @@
     </style>
 
     <style name="Theme.CapstoneMD" parent="Base.Theme.CapstoneMD" />
+
+    <style name="Theme.CapstoneMD.NoActionBar">
+        <item name="windowActionBar">false</item>
+        <item name="windowNoTitle">true</item>
+    </style>
+
+    <style name="Theme.CapstoneMD.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+
+    <style name="Theme.CapstoneMD.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 </resources>
